prevent duplicate password requests
